requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
gunicorn>=21.2
sse-starlette>=2.1
jinja2>=3.1
python-multipart>=0.0.9
pydantic>=2.5
pydantic-settings>=2.2
pydicom>=2.4
opencv-python-headless>=4.9
numpy>=1.26
pillow>=10.0
httpx>=0.27
weasyprint>=60.2
python-docx>=1.1
itsdangerous>=2.2
# ML deps match legacy echochat env; align to /nfs/usrhome2/yqinar/HKUSData/script/echochatv1.5
ms-swift>=3.3
torch>=2.1
transformers>=4.51
# Dev
pytest>=8.0
pytest-asyncio>=0.23
respx>=0.21
